requirements.txt
openai
python-dotenv
Pillow
boto3
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client